# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Youtube Video API",
        "description": "This is a Youtube Video API that allow users to retrieve generated data about youtube video(s) as a JSON format.",
        "contact": {
            "email": "n00201327@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register account",
                "description": "Register an account to retrieve data from API",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "description": "Login data",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter your channel name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Enter your channel email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Enter your channel password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully registered an account.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated: Email is already used by another user."
                    },
                    "406": {
                        "description": "Unacceptable: Missing some or all parameters."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to your account",
                "description": "Login to your account to retrieve data from API",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "description": "Login data",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Enter your channel email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Enter your channel password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/channel": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get details about your channel",
                "description": "Get everything related to your account such as name, subscriber, videos, comments etc.",
                "operationId": "5f0516e37fc40ea57a3d8e0f87462544",
                "responses": {
                    "200": {
                        "description": "Successful operation returns JSON format",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Channel"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Comment"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update all details in your channel",
                "description": "Update Channel details",
                "operationId": "44410b6f2543a98ffd159c723e86b5bd",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Update your channel name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Update your channel email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Update your channel password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Delete your channel",
                "description": "Delete your own channel from the database, including the videos, and comments by the channel",
                "operationId": "9c327252e6821d149ce77010c2dced64",
                "responses": {
                    "200": {
                        "description": "Channel has been successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update a specific detail in your channel",
                "description": "Update Channel details",
                "operationId": "8ca3e06b311fdf839c2a949473dc3b87",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Update your channel name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Update your channel email",
                                        "type": "email"
                                    },
                                    "password": {
                                        "description": "Update your channel password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Channel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout of your account",
                "description": "Logout of your account.",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Channel has been successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/auth"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "youtube_video": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit."
                    },
                    "description": {
                        "type": "text",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque sed erat mi. Vivamus mollis orci vitae neque fringilla, non interdum urna commodo. Cras at blandit lorem, malesuada posuere felis."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-10-29"
                    },
                    "duration": {
                        "type": "integer",
                        "example": "30"
                    },
                    "likes": {
                        "type": "integer",
                        "example": "30"
                    },
                    "dislikes": {
                        "type": "integer",
                        "example": "30"
                    },
                    "views": {
                        "type": "integer",
                        "example": "30"
                    },
                    "link": {
                        "type": "text",
                        "example": "youtube.com\\/watch?v=1fa3d3b7-d0c7-3e5e-89e7-3713d0e74f87"
                    },
                    "thumbnail": {
                        "type": "text",
                        "example": "https:\\/\\/via.placeholder.com\\/640x480.png\\/005500?text=thumbnail+voluptatem"
                    },
                    "channel": {
                        "$ref": "#/components/schemas/Channel"
                    },
                    "comments": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object"
            },
            "Channel": {
                "properties": {
                    "id": {
                        "description": "channel id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "name of the channel",
                        "type": "string",
                        "example": "James"
                    },
                    "subscribers": {
                        "description": "number of subscribers",
                        "type": "integer",
                        "example": "1100"
                    },
                    "created_at": {
                        "description": "the date when the channel has been created",
                        "type": "string",
                        "format": "date",
                        "example": "2019-10-06"
                    },
                    "hasVideo": {
                        "description": "the amount of video",
                        "type": "integer",
                        "format": "integer",
                        "example": "100"
                    },
                    "videos": {
                        "$ref": "#/components/schemas/youtube_video"
                    }
                },
                "type": "object"
            },
...
            "auth": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "201"
                    },
                    "message": {
                        "type": "string",
                        "example": "Channel has been successfully registered"
                    },
                    "token": {
                        "type": "string",
                        "example": "1|C7eINx2yy8TUM3PkWZXwt8OtN62ndluhpcYKjxIE"
                    }
                },
                "type": "object"
            },
            "Comment": {
                "properties": {
                    "id": {
                        "description": "comment id",
                        "type": "integer",
                        "example": "1"
                    },
                    "text": {
                        "description": "comment texts",
                        "type": "text",
                        "example": "lorem ipsum"
                    },
                    "likes": {
                        "description": "number of likes",
                        "type": "integer",
                        "example": "1100"
                    },
                    "dislikes": {
                        "description": "number of dislikes",
                        "type": "integer",
                        "example": "90"
                    },
                    "commented_at": {
                        "description": "The date of the comment",
                        "type": "string",
                        "format": "string",
                        "example": "6 years ago"
                    },
                    "channel": {
                        "$ref": "#/components/schemas/Channel"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}